Fix "unsafe:" prefix in anchor links
Along with AngularJS 1.7, javascript-protocol links (e.g. "javascript:void(0)" were prefixed with "unsafe:", which causes a page navigation in Safari. This commit sets the href attributes to be empty for the following reasons:

* Adding "#" results in navigation being detected by the router.

* Changing <a> to <button> elements makes too many assumptions when it comes to styling and causes UI issues.

Note: Instead of this patch, the javascript "protocol" could be added to the sanitation whitelist, which on the other hand side might result in security relevant side effects:

```
$compileProvider.aHrefSanitizationWhitelist(/^\s*(https?|s?ftp|mailto|tel|file|javascript):/);
```

Resolves: https://github.com/720kb/angular-datepicker/issues/318

This bumps to version: 2.1.7-itonics